requirements.txt
Mastodon.py==1.8.1
requests==2.30.0